Add Dockerfile and .dockerignore to text and image, move affiliates.js into shared
