+++ shared/affiliates.js
