+++ text.pollinations.ai/package.json
@@ -30,7 +30,9 @@
         "openai": "^4.58.2",
         "p-queue": "^8.0.1",
         "puppeteer": "^23.6.0",
-        "turndown": "^7.2.0"
+        "turndown": "^7.2.0",
+        "tsx": "^4.20.3",
+        "typescript": "^5.8.3"
     },
     "devDependencies": {
         "@istanbuljs/nyc-config-babel": "^3.0.0",
@@ -39,8 +41,6 @@
         "nodemon": "^3.1.10",
         "nyc": "^17.1.0",
         "sinon": "^20.0.0",
-        "supertest": "^7.0.0",
-        "tsx": "^4.20.3",
-        "typescript": "^5.8.3"
+        "supertest": "^7.0.0"
     }
 }
